kas/opt: Add option for not install kernel
This adds an option to create an image that does not contain the kernel.
This is mainly intended to provide images for kernel testing.
When testing the kernel, kernel image files that are not used for testing
are not necessary in rootfs. Providing images with these files removed
allows you to reduce the size of the rootfs image.

wo_kernel is false:
  -rw-r--r-- 1 iwamatsu iwamatsu 351M Jul 25 17:31 cip-core-image-cip-core-bookworm-qemu-amd64.ext4

no_kernel is true:
  -rw-r--r-- 1 iwamatsu iwamatsu 275M Jul 25 17:09 cip-core-image-cip-core-bookworm-qemu-amd64.ext4

Signed-off-by: Nobuhiro Iwamatsu <nobuhiro1.iwamatsu@toshiba.co.jp>
[Jan: update all call sites of deploy-cip-core.sh]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>