erofs: Set UUID and SOURCE_DATE_EPOCH with empty exclude dir
if EROFS_EXCLUDE_DIR is empty then SOURCE_DATE and UUID are not
set. These variables should not be depend on EROFS_EXCLUDE_DIR
as the are necessary for reproducible erofs builds.

Also add a space to setting the uuid string. This avoids
the mangle of EROFS_CREATION_ARGS in case EROFS_EXCLUDE_DIRS
is empty.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>