make apt cache world readable
The apt cache directory in /var/lib/apt has default permissions of 755
on non-ro systems. By that, unprivileged users (or tooling) can access
the apt cache. When running on a ro-rootfs, we need to create the
directory using systemd tmpfiles, as the system is deployed with an
empty /var. This should be done with the same permissions as before.

Note, that this should not have any effect on the target systems, as
we anyways wipe the apt cache before deploying. However, as this
tmpfiles snippet comes as a package, the limitation already applies at
assembly time, which breaks BOM scanners that run as a rootfs
postprocess command.

Fixes: 26763b3 ("create apt directory as needed by apt-daily")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>