linux-cip: Disable CONFIG_HW_RANDOM_TPM in QEMU
It turned out that the fTPM at least writes every 4 seconds its time to
flash when executing commands while hwrng asks it for a random number
every minute if this config is enabled, triggering the write.

Disabling is recommended for now, also for discrete TPMs. We do that via
config snippet so that the official CIP kernel config keeps the feature
in scope.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>