qemu-arm64: Add optee-os 4.8.0
Build OP-TEE for qemu-arm64 with fTPM and StMM. If QEMU comes with an
RPMB emulation, this allows to run a firmware stack that is very close
to real devices. It specifically enables modifiable secure UEFI
variables and allows to drop the virtual discrete TPM.

For building the fTPM TA, also the corresponding tadevkit package is
generated.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>