kas: Add option for fTPM/StMM firmware mode
This switches the qemu-arm64 target to TF-A, OP-TEE, fTPM and StMM
based firmware. We only support this with kernel 6.12 due to features
only available from that version onward. Furthermore the toolchain of
buster is not supported by optee-ftpm, so exclude this variant.

As this feature currently requires a custom QEMU build, do not turn it
on by default.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>