.reproducible-check-ci.yml: Use S3 instead of the gitlab artifact storage
So far we have been using the gitlab artifact storage for the build artifacts
from the RB tests. But this has the max artifact size per job set to 1 GB and
lately our builds seem to slightly go over this mark.

Let's use S3 to store the build artifacts and later fetch from the same during
the test stage.

Signed-off-by: Adithya Balakumar <Adithya.Balakumar@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>