linux-cip: Refactor KERNEL_DEFCONFIG initialization
Now that all targets are on cip_merged_defconfig, we can make this the
standard value of KERNEL_DEFCONFIG if USE_CIP_KERNEL_CONFIG is set.
This is logically best done in cip-kernel-config.inc.

While at it, switch to bb.utils.to_boolean for the evaluation of
USE_CIP_KERNEL_CONFIG.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>