Add DE0-Nano-SoC machine
We boot the board from SD card in EFI mode, with or without SWUpdate
support.

For that, we need a hybrid partition table. The entry containing the
firmware is marked with the type 0xa2 in the legacy msdos partition
table. Unfortunately, this cannot be expressed with wic natively because
it will try to apply --system-id to the GPT, failing to do so with msdos
value. Therefore add a custom bootloader plugin that does that
hard-coded at the end of the image generation.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>