IEC_template.yml: Update IEC layer tests revision
This revision update fixes the inconsistency with the IEC CR 7.1 test.

Currently the rate limit using nftables is set to 3 requests/second and
the ping flood transmits a packet every 0.3 seconds i.e 3.3 packets / second

Although this is theoretically over the rate limit and should drop packets,
but the rate limit set and the ping requests interval is too close and some
iterations can lead to no packets being dropped and thereby the test failing.

The fix increase the frequency in the ping flood attack which would lead to
more packets being dropped by the kernel.

Signed-off-by: Adithya Balakumar <Adithya.Balakumar@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>