Add factory-reset-helper to all images
Currently, the factory-reset-helper package is conditionally appended to
IMAGE_INSTALL only in cip-core-image.bb

But the security image uses cip-core-image-security.bb as the top-level
image recipe and therefore does not include the factory-reset-helper package.

Let's move appending factory-reset-helper to IMAGE_INSTALL to a common place
so that its also available to all images.

Closes #141

Signed-off-by: Adithya Balakumar <Adithya.Balakumar@toshiba-tsip.com>
[Jan: generalized commit message]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>