Update isar revision
Brings some smaller fixes of which we specifically need the one for the
ftpm hook to ensure reliable booting on the TI AM62Px.

One API change comes with this version: INITRD_IMAGE was deprecated, and
the new IMAGE_INITRD is takes now INITRAMFS_RECIPE directly. That change
came with several issues, the critical ones are temporarily addressed by
two upstream-pending patches.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>