u-boot-ti-am62px-sk: ti-extra.cfg: Remove redundant configs
TEE, OPTEE, CMD_OPTEE_RPMB and EFI_MM_COMM_TEE are enabled in
ftpm-stmm.cfg and it is always included for TI AM62Px builds. So remove
it from ti-extra.cfg.

The BOOTCOMMAND is stale which was initially used for testing purposes.
pairing with RPMB is not recommended in production environments and also
CFG_RPMB_WRITE_KEY=y is not set so this doesn't get triggered. Remove
it.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>