start-qemu.sh: Check FTPM_STMM instead of TPM2_DEVICE before setting RPMB options for arm64
Current script checks for the absence of TPM2_DEVICE to set RPMB options
for QEMU arm64. TPM2_DEVICE may not be set for a SWUPDATE only boot as well and
the script looks for the RPMB image even if FTPM_STMM is not set.

Instead, let's check if FTPM_STMM is set.

Signed-off-by: Adithya Balakumar <Adithya.Balakumar@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>